requirements.txt
streamlit
langchain-community
langchain-core
langchain-google-genai
langchain-huggingface
sentence-transformers
faiss-cpu
pypdf2
python-dotenv